Move isFree checkbox to search filters
