--- src/domain/eventSearch/searchResultList/extraFilters/extraFilters.module.scss
@@ -1,18 +1,0 @@
-@import "colors";
-@import "fonts";
-@import "layout";
-@import "variables";
-
-.extraFilters {
-  margin-top: $largeMargin;
-
-  .contentWrapper {
-    padding: $basePadding;
-    background-color: $hel-white;
-
-    .title {
-      font-size: 1.125rem;
-      margin: 0 0 0.875rem;
-    }
-  }
-}
